BLD: Reduce the number of Travis build permutations.
Now that the pandas and numpy upgrade is complete, no longer need to
test for forwards and backwards compatibility.

Also, now that Cython is part of the build, the existence of Cython can
be assumed for building cyordereddict.

Goal is to cut down the number of permutationts to just Python 2 and 3,
for a faster CI cycle.